more reorg things
- made a serebii.parse_util file for serebii specific things that were in the general util file (hopefully I fixed all the imports???)
- moved all pokeapi-specific logic from forms into new pokeapi.parser file
- removed the form config since I'm going to rewrite a new one
