+++ scripts/compare/abilityscript.py
@@ -4,7 +4,7 @@
 import requests
 from lxml import html
 
-from scripts.util import add_row_values
+from scripts.serebii.parse_util import add_row_values
 
 f = open("abilities.in", "r")
 out = open("abilities.out", "w")
